requirements.txt
kafka-python~=2.2.4
pylint~=3.3.1
pyyaml~=6.0.2
